backend/requirements.txt
fastapi==0.112.0
starlette==0.37.2
uvicorn==0.30.5
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.1.0
ultralytics==8.3.70
pydantic==2.10.4
python-dotenv==1.0.1
cloudinary==1.42.1
openai==1.60.2
python-multipart==0.0.20